Extract `SaveConfig` from `RotateConfig`.
