+++ src/app-config/export/RotateConfig.js
@@ -3,15 +3,16 @@
 
 import { MustBe } from '@this/typey';
 
-import { BaseConfig } from '#x/BaseConfig';
+import { SaveConfig } from '#x/SaveConfig';
 
 
 /**
  * Configuration representation for file rotation, used in configuring some
- * file-writing services. (See {@link FileServerConfig}.)
+ * file-writing services. (See {@link FileServiceConfig}.)
  *
  * Accepted configuration bindings (in the constructor). All are optional.
  *
+ * * Everything accepted by {@link RotateConfig}.
  * * `{?number} atSize` -- Rotate when the file becomes the given size (in
  *   bytes) or greater. If `null`, does not rotate based on size. Default
  *  `null`.
@@ -21,17 +22,8 @@ import { BaseConfig } from '#x/BaseConfig';
  * * `{?number} maxOldBytes` -- How many bytes' worth of old (post-rotation)
  *   files should be allowed, or `null` not to have a limit. The oldest files
  *   over the limit get deleted after a rotation. Default `null`.
- * * `{?number} maxOldCount` -- How many old (post-rotation) files should be
- *   allowed, or `null` not to have a limit. The oldest files over the limit get
- *   deleted after a rotation. Default `null`.
- * * `{?boolean} onReload` -- Rotate when the system is reloaded (restarted
- *   in-process). Default `false`.
- * * `{?boolean} onStart` -- Rotate when the system is first started? Default
- *   `false`.
- * * `{?boolean} onStop` -- Rotate when the system is about to be stopped?
- *   Default `false`.
  */
-export class RotateConfig extends BaseConfig {
+export class RotateConfig extends SaveConfig {
   /** @type {?number} The file size at which to rotate, if ever. */
   #atSize;
 
@@ -48,20 +40,6 @@ export class RotateConfig extends BaseConfig {
   #maxOldBytes;
 
   /**
-   * @type {?number} The maximum number of old files to allow, if so limited.
-   */
-  #maxOldCount;
-
-  /** @type {boolean} Rotate when reloading the system? */
-  #onReload;
-
-  /** @type {boolean} Rotate when starting the system? */
-  #onStart;
-
-  /** @type {boolean} Rotate when stopping the system? */
-  #onStop;
-
-  /**
    * Constructs an instance.
    *
    * @param {object} config Configuration object. See class header for details.
@@ -72,11 +50,7 @@ export class RotateConfig extends BaseConfig {
     const {
       atSize      = null,
       checkSecs   = 5 * 60,
-      maxOldBytes = null,
-      maxOldCount = null,
-      onReload    = false,
-      onStart     = false,
-      onStop      = false,
+      maxOldBytes = null
     } = config;
 
     this.#atSize = (atSize === null)
@@ -86,12 +60,6 @@ export class RotateConfig extends BaseConfig {
     this.#maxOldBytes = (maxOldBytes === null)
       ? null
       : MustBe.number(maxOldBytes, { finite: true, minInclusive: 1 });
-    this.#maxOldCount = (maxOldCount === null)
-      ? null
-      : MustBe.number(maxOldCount, { finite: true, minInclusive: 1 });
-    this.#onReload = MustBe.boolean(onReload);
-    this.#onStart  = MustBe.boolean(onStart);
-    this.#onStop   = MustBe.boolean(onStop);
 
     if (this.#atSize === null) {
       // `checkSecs` is irrelevant in this case.
@@ -121,28 +89,5 @@ export class RotateConfig extends BaseConfig {
    */
   get maxOldBytes() {
     return this.#maxOldBytes;
-  }
-
-  /**
-   * @returns {?number} The maximum number of old files to allow, or `null` if
-   * there is no limit.
-   */
-  get maxOldCount() {
-    return this.#maxOldCount;
-  }
-
-  /** @returns {boolean} Rotate when reloading the system? */
-  get onReload() {
-    return this.#onReload;
-  }
-
-  /** @returns {boolean} Rotate when starting the system? */
-  get onStart() {
-    return this.#onStart;
-  }
-
-  /** @returns {boolean} Rotate when stopping the system? */
-  get onStop() {
-    return this.#onStop;
   }
 }
