+++ src/app-config/index.js
@@ -14,6 +14,7 @@ export * from '#x/MountConfig';
 export * from '#x/NamedConfig';
 export * from '#x/Names';
 export * from '#x/RotateConfig';
+export * from '#x/SaveConfig';
 export * from '#x/ServiceConfig';
 export * from '#x/Uris';
 export * from '#x/Util';
